refactoring main

+++ src/image/image.h
@@ -21,26 +21,14 @@ typedef struct image
 {
 	mlx_t		*win;
 	mlx_image_t	*img;
-	size_t		width;
-	size_t		height;
+	int32_t		width;
+	int32_t		height;
 	char		*title;
 	bool		resizeable;
 }				t_image;
 
-t_image *
-create_window(
-	size_t w,
-	size_t h,
-	char *title,
-	bool resize);
-
-void			set_pixel(
-	t_image *image,
-	t_color color,
-	int i,
-	int j);
-
-void			draw_loop(
-	mlx_t *win);
+t_image	*create_window(size_t w, size_t h, char *title, bool resize);
+void	set_pixel(t_image *image, t_color color, int i, int j);
+void	draw_loop(mlx_t *win);
 
 #endif
